Paused work before removing the Large Vector struct
